% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_gbcd.R
\name{fit_gbcd}
\alias{fit_gbcd}
\alias{fit_gbcd_control_default}
\title{Fit Generalized Binary Covariance Decomposition}
\usage{
fit_gbcd(
  Y,
  Kmax,
  prior = ebnm::ebnm_generalized_binary,
  maxiter1 = 500,
  maxiter2 = 200,
  maxiter3 = 500,
  control = list(),
  verbose = 1
)

fit_gbcd_control_default()
}
\arguments{
\item{Y}{Cell x gene matrix of normalized and log-transformed gene
expression data.}

\item{Kmax}{Integer 2 or greater used to determine the number of
final number of factors in the GBCD. The final number of factors is
often close to \code{Kmax}, and is never greater than \code{2*Kmax
- 1}.}

\item{prior}{Nonnegative prior for GEP memberships, usually the
generalized binary prior. The ratio \eqn{\rho = \sigma_k/\mu_k} in
the generalized binary prior can be controlled via the
\code{\link[flashier]{flash_ebnm}} function from flashier,
\code{flash_ebnm(prior_family = "generalized_binary", scale =
rho)}. See also the \code{\link[ebnm]{ebnm_generalized_binary}}
function from the ebnm package for more information.}

\item{maxiter1}{Positive integer specifying the maximum number of
backfit iterations during the GEP membership matrix L
initialization.}

\item{maxiter2}{Positive integer specifying the maximum number of
backfit iterations during the GEP membership matrix L estimation.}

\item{maxiter3}{Positive integer specifying the maximum number of
backfit iterations during the GEP signature matrix F estimation.}

\item{control}{List of control parameters with the following
elements: \dQuote{warmstart}, a logical indicator specifying
whether to use warmstart to initialize the prior g when solving
EBNM subproblems (see ebnm for details); \dQuote{extrapolate}, a
logical indicator specifying whether to use extrapolation to
accelerate backfitting GEP memberships (see flashier for details);
\dQuote{corr_thres}, a numeric scalar between 0 and 1 such that we
only keep columns \eqn{l_k} whose Pearson correlation with
\eqn{\tilde{l}_k} exceeds \dQuote{corr_thres}.}

\item{verbose}{Integer specifying whether and how to display
progress updates, as described in flashier.}
}
\value{
A list including the following elements:

\item{L}{cell x GEP matrix containing the posterior estimates of
the GEP membership matrix L.}

\item{F}{List containing the posterior summaries of the GEP
  signature matrix F. Note that these posterior estimates are
  provided on the base-2 logarithmic scale for more straightforward
  interpretation as \emph{log-fold changes}; to return these estimates
  to the usual (natural log) scale, you would multiply these by
  \code{log(2)}. For example, fitted values would be recovered as
  \code{tcrossprod(L,log(2) * F$lfc)}.}
}
\description{
Fit a generalized binary covariance decomposition (GBCD)
to single cell RNA-seq data containing multiple tumors.
}
\examples{
# Please see the vignettes for examples.

}
